Use cross-spawn to diff
